chore: move example tools
