+++ examples/agents/tools/tools.ts
