+++ examples/deprecated/agents/openai-task.ts
@@ -4,7 +4,7 @@ import {
   getCurrentIDTool,
   getUserInfoTool,
   getWeatherTool,
-} from "./utils/tools";
+} from "../../agents/tools/tools";
 
 async function main() {
   // Create an OpenAIAgent with the function tools
